requirements.txt
pyqt5>=5.12
requests>=2.18
colorama>=0.3.8
pyyaml>=3.13
